requirements.txt
wheel
py-postgresql
requests
tibber.py
